New Template for StarlingX Specs
This document is intended to form the basis for new specs for the
StarlingX project.  The template is based on the Stein Nova Spec
template, which has been refined over time, and provides an
excellent foundation. The template is long, but, there are optional
sections.Overall, this captures the key elements of a spec that will
help the project define new technology directions.

Change-Id: Icf5177289ca4001dad0057e90000ddfaf6cce051
Signed-off-by: Ian Jolliffe <ian.jolliffe@windriver.com>