Spec for multiOS architecture
Starling X is based on CentOS operating system, however, is necessary to give
support of the product in other base OS layers. The current build system under
STX will not support the transition to other OS based on rpm or deb files. In
order to give a full support for other operating systems is necessary to
refactor the current build model and generate an isolation of source code of
stx flock services, patches, and build tools

Change-Id: Ic8511d06dfb08daba5e96d49465bac9d1d43e6d5
Signed-off-by: VictorRodriguez <vm.rod25@gmail.com>