Multi OS Linux builder tool
Starling X is based on CentOS operating system, however, is necessary to give
support of the product in other base OS layers. The current build system under
STX will not support the transition to other OS based on rpm or deb files. In
order to give a full support for other operating systems is necessary to create
a tool that make linux images based on deb files (like ubuntu) and in rpms from
other Linux distributions ( like fedora )

Change-Id: I36d3efbd3c2f7c033e528b60a5b912dfa627a44a
Signed-off-by: VictorRodriguez <vm.rod25@gmail.com>