Debian base image build tool (POC)
Starling X is based on CentOS operating system, however, in order to increase
the adoption of the project, is necessary to support the build of STX in other
OS flavors. This specification provide a POC of a new build image tool to
create ubuntu images with specific deb files. How the DEB is built is out of
the scope of this tool

Change-Id: I36d3efbd3c2f7c033e528b60a5b912dfa627a44a
Signed-off-by: VictorRodriguez <vm.rod25@gmail.com>