Documentation Mega Spec
This specification describes a major change for the StarlingX
documentation (stx.docs / docs.starlingx.io) designed to address
use cases for users, operators and contributors.  It proposes
both a refactoring of the existing documentation and the creation
of a large number of additional documents.  A large number of
StoryBoard stories will be created from the proposal in this
document, which will all be prioritized as described in the spec.

I am pleased to submit this spec for review on behalf of the Docs
team and would like to thank Michael, Kris, Scott, Abraham and
Greg for their many contributions to this spec.

Change-Id: I48e3728e576ca72f00c77a049d5e5d8a229ab3de
Signed-off-by: Bruce Jones <bruce.e.jones@intel.com>