Reorganize Flock Services Source Code repositories
StarlingX currently combines source code and build meta-data in the same
directory structure, as we move to support multiple OSes, this structure needs
to be separated and further refined for the benefit of developers in the
community to keep history and source code separated. Simplifying things to a
set of common features also has the dual benefit of making the code structure
easier to understand and contribute to.

Change-Id: I6ffe451ef48796411b205bc3f84bd485081503db
Signed-off-by: VictorRodriguez <vm.rod25@gmail.com>
Suggested-by: Saul Wold <saul.wold@intel.com>