[WIP] StarlingX Linux Kernel
Posting to get preliminary feedback on the scope of this spec.

StarlingX [0] relies on the Linux kernel to manage the hardware, run user
programs, and maintain the security and integrity of the whole system. As a
fully featured cloud project for the distributed edge, there is a commitment
to support emerging requirements, compute intensive use cases and workloads,
which will make become a reality the imperatives of quality of service,
performance, security and low latency...

Change-Id: I8d19605274d952b0a422b6d09c9575899c39cffa
Signed-off-by: Abraham Arce <abraham.arce.moreno@intel.com>